+++ Runtime/Scripts/GltfImport.cs
@@ -2013,6 +2013,9 @@ namespace GLTFast {
                     if (att.TEXCOORD_1 >= 0) {
                         uvInputs[1] = att.TEXCOORD_1;
                     }
+                    if (att.TEXCOORD_2 >= 0) {
+                        logger.Warning(LogCode.UVLimit);
+                    }
                 }
 
                 VertexBufferConfigBase config;